[svn r11] New db parser and better utility module.
The new parser that loads a matrix now relies on the information generated by
the Python analyzer so a single pass is sufficient to load the information.
Thanks to this dependence, the parser can be greatly simplified as most of the
syntax errors will be detected by the Python script.

Along with this change, the util module has been enhanced with a wider and
better variety of messages, as well as robust versions of memory allocation
primitives.

The rest of the changes involved a bit of reorganization and modularization.