[svn r14] Unified treatment of the data files.
It has been revealed that sometimes there is a mismatch between the dimensions
of the training data and the test data.  The solution is to treat both files at
once: at the analyzer and the benchmarker.

Therefore, and in both cases, there's only required to specify the test data
file for treatment.  Then the training data is assumed to be in a file with the
same name and the ".t" extension appended.

Info files are still created separately but some information is joined before
creating them.