[svn r74] Fixed the last standing bogus blocking implementation.
The unroll4 version, in the blocking vectorized with bytes implementation
revealed an issue with the block size specific to the unrolled versions.  A new
function has been added to further adjust the block item count, so no residual
iterations are performed in any block but the last.

They have been added to the nnsca-*.c files, as they have fewer lines than their
corresponding nnvec-*.c