Dump unblocked runs from benchmarking and plots.
After proving the stability and reliability of "chunking" in order to raise the
cache hit ratio, the number of benchmark runs can be cut in half by removing
"unblocked" timings (as they are no longer interesting).

The same happens when plotting the results.

The block size has been reduced slightly because it seems it may improve
performance a little.