Simplify OpenCL specific error printing.
Thanks to the macro, error locations are easily spotted so there is no need to
be creative in an error message for each different OpenCL call.

Therefore, the ocl_error and ocl_fatal macros have been changed to include the
error check inside them.  Thus cleaning a little bit real code, and also giving
the chance of placing them in the same line as the affected call, so the line
information is even more precise.